+++ dynamo/preprocessing/deprecated.py
@@ -40,8 +40,8 @@ from .QC import (
     filter_cells_by_outliers,
     filter_genes_by_outliers,
 )
+from .transform import _Freeman_Tukey
 from .utils import (
-    _Freeman_Tukey,
     add_noise_to_duplicates,
     calc_new_to_total_ratio,
     collapse_species_adata,
