+++ dynamo/preprocessing/preprocessor_utils.py
@@ -28,11 +28,9 @@ from ..dynamo_logger import (
     main_log_time,
     main_warning,
 )
-from ..tools.utils import update_dict
 from ..utils import copy_adata
 from .pca import pca
 from .utils import (
-    _Freeman_Tukey,
     add_noise_to_duplicates,
     calc_new_to_total_ratio,
     collapse_species_adata,
@@ -41,25 +39,6 @@ from .utils import (
     get_svr_filter,
     merge_adata_attrs,
 )
-
-
-def is_log1p_transformed_adata(adata: anndata.AnnData) -> bool:
-    """check if adata data is log transformed by checking a small subset of adata observations.
-
-    Args:
-        adata: an AnnData object
-
-    Returns:
-        A flag shows whether the adata object is log transformed.
-    """
-
-    chosen_gene_indices = np.random.choice(adata.n_vars, 10)
-    _has_log1p_transformed = not np.allclose(
-        np.array(adata.X[:, chosen_gene_indices].sum(1)),
-        np.array(adata.layers["spliced"][:, chosen_gene_indices].sum(1)),
-        atol=1e-4,
-    )
-    return _has_log1p_transformed
 
 
 def _infer_labeling_experiment_type(adata: anndata.AnnData, tkey: str) -> Literal["one-shot", "kin", "deg"]:
@@ -167,299 +146,6 @@ def seurat_get_mean_var(
     mean = np.nan_to_num(mean)
     var = np.nan_to_num(var)
     return mean, var
-
-
-def log1p_adata_layer(adata: AnnData, layer: str = DKM.X_LAYER, copy: bool = False) -> AnnData:
-    """Calculate log1p of adata's specific layer.
-
-    Args:
-        adata: an AnnData object.
-        layer: the layer to operate on. Defaults to DKM.X_LAYER.
-        copy: whether operate on the original object or on a copied one and return it. Defaults to False.
-
-    Returns:
-        The updated AnnData object.
-    """
-
-    _adata = adata
-    if copy:
-        _adata = copy_adata(adata)
-    log1p_inplace(_adata, layer=layer)
-    return _adata
-
-
-def log2_adata_layer(adata: AnnData, layer: str = DKM.X_LAYER, copy: bool = False) -> AnnData:
-    """Calculate log2 of adata's specific layer.
-
-    Args:
-        adata: an AnnData object.
-        layer: the layer to operate on. Defaults to DKM.X_LAYER.
-        copy: whether operate on the original object or on a copied one and return it. Defaults to False.
-
-    Returns:
-        The updated AnnData object.
-    """
-
-    _adata = adata
-    if copy:
-        _adata = copy_adata(adata)
-    log2_inplace(_adata, layer=layer)
-    return _adata
-
-
-def log_adata_layer(adata: AnnData, layer: str = DKM.X_LAYER, copy: bool = False) -> AnnData:
-    """Calculate log of adata's specific layer.
-
-    Args:
-        adata: an AnnData object.
-        layer: the layer to operate on. Defaults to DKM.X_LAYER.
-        copy: whether operate on the original object or on a copied one and return it. Defaults to False.
-
-    Returns:
-        The updated AnnData object.
-    """
-
-    _adata = adata
-    if copy:
-        _adata = copy_adata(adata)
-    log_inplace(_adata, layer=layer)
-    return _adata
-
-
-def Freeman_Tukey_adata_layer(adata: AnnData, layer: str = DKM.X_LAYER, copy: bool = False) -> AnnData:
-    """Calculate Freeman_Tukey of adata's specific layer.
-
-    Args:
-        adata: an AnnData object.
-        layer: the layer to operate on. Defaults to DKM.X_LAYER.
-        copy: whether operate on the original object or on a copied one and return it. Defaults to False.
-
-    Returns:
-        The updated AnnData object.
-    """
-
-    _adata = adata
-    if copy:
-        _adata = copy_adata(adata)
-    Freeman_Tukey_inplace(_adata, layer=layer)
-    return _adata
-
-
-def log1p(adata: AnnData, layers: list = [DKM.X_LAYER], copy: bool = False) -> AnnData:
-    """Perform log1p transform on selected adata layers
-
-    Args:
-        adata: an AnnData object.
-        layers: the layers to operate on. Defaults to [DKM.X_LAYER].
-        copy: whether operate on the original object or on a copied one and return it. Defaults to False.
-
-    Returns:
-        The updated AnnData object.
-    """
-
-    _adata = adata
-    if copy:
-        _adata = copy_adata(adata)
-
-    main_debug("[log1p] transform applied to layers: %s" % (str(layers)))
-    for layer in layers:
-        log1p_adata_layer(_adata, layer=layer)
-
-    main_info_insert_adata_uns("pp.norm_method")
-    adata.uns["pp"]["norm_method"] = "log1p"
-    return _adata
-
-
-def log2(adata: AnnData, layers: list = [DKM.X_LAYER], copy: bool = False) -> AnnData:
-    """Perform log2 transform on selected adata layers
-
-    Args:
-        adata: an AnnData object.
-        layers: the layers to operate on. Defaults to [DKM.X_LAYER].
-        copy: whether operate on the original object or on a copied one and return it. Defaults to False.
-
-    Returns:
-        The updated AnnData object.
-    """
-
-    _adata = adata
-    if copy:
-        _adata = copy_adata(adata)
-
-    main_debug("[log2] transform applied to layers: %s" % (str(layers)))
-    for layer in layers:
-        log2_adata_layer(_adata, layer=layer)
-
-    main_info_insert_adata_uns("pp.norm_method")
-    adata.uns["pp"]["norm_method"] = "log2"
-    return _adata
-
-
-def log(adata: AnnData, layers: list = [DKM.X_LAYER], copy: bool = False) -> AnnData:
-    """Perform log transform on selected adata layers
-
-    Args:
-        adata: an AnnData object.
-        layers: the layers to operate on. Defaults to [DKM.X_LAYER].
-        copy: whether operate on the original object or on a copied one and return it. Defaults to False.
-
-    Returns:
-        The updated AnnData object.
-    """
-
-    _adata = adata
-    if copy:
-        _adata = copy_adata(adata)
-
-    main_debug("[log] transform applied to layers: %s" % (str(layers)))
-    for layer in layers:
-        log_adata_layer(_adata, layer=layer)
-
-    main_info_insert_adata_uns("pp.norm_method")
-    adata.uns["pp"]["norm_method"] = "log"
-    return _adata
-
-
-def Freeman_Tukey(adata: AnnData, layers: list = [DKM.X_LAYER], copy: bool = False) -> AnnData:
-    """Perform Freeman_Tukey transform on selected adata layers
-
-    Args:
-        adata: an AnnData object.
-        layers: the layers to operate on. Defaults to [DKM.X_LAYER].
-        copy: whether operate on the original object or on a copied one and return it. Defaults to False.
-
-    Returns:
-        The updated AnnData object.
-    """
-
-    _adata = adata
-    if copy:
-        _adata = copy_adata(adata)
-
-    main_debug("[Freeman_Tukey] transform applied to layers: %s" % (str(layers)))
-    for layer in layers:
-        Freeman_Tukey_adata_layer(_adata, layer=layer)
-
-    main_info_insert_adata_uns("pp.norm_method")
-    adata.uns["pp"]["norm_method"] = "Freeman_Tukey"
-    return _adata
-
-
-def _log1p_inplace(data: np.ndarray) -> np.ndarray:
-    """Calculate log1p (log(1+x)) of an array and update the array inplace.
-
-    Args:
-        data: the array for calculation.
-
-    Returns:
-        The updated array.
-    """
-
-    return np.log1p(data, out=data)
-
-
-def _log2_inplace(data: np.ndarray) -> np.ndarray:
-    """Calculate Base-2 logarithm of `x` of an array and update the array inplace.
-
-    Args:
-        data: the array for calculation.
-
-    Returns:
-        The updated array.
-    """
-
-    return np.log2(data + 1, out=data)
-
-
-def _log_inplace(data: np.ndarray) -> np.ndarray:
-    """Calculate the natural logarithm `log(exp(x)) = x` of an array and update the array inplace.
-
-    Args:
-        data: the array for calculation.
-
-    Returns:
-        The updated array.
-    """
-
-    return np.log(data + 1, out=data)
-
-
-def log1p_inplace(adata: AnnData, layer: str = DKM.X_LAYER) -> None:
-    """Calculate log1p (log(1+x)) for a layer of an AnnData object inplace.
-
-    Args:
-        adata: an AnnData object.
-        layer: the layer to operate on. Defaults to DKM.X_LAYER.
-    """
-
-    mat = DKM.select_layer_data(adata, layer, copy=False)
-    if issparse(mat):
-        if is_integer_arr(mat.data):
-            mat = mat.asfptype()
-            DKM.set_layer_data(adata, layer, mat)
-        _log1p_inplace(mat.data)
-    else:
-        mat = mat.astype(np.float)
-        _log1p_inplace(mat)
-
-
-def log2_inplace(adata: AnnData, layer: str = DKM.X_LAYER) -> None:
-    """Calculate Base-2 logarithm of `x` for a layer of an AnnData object inplace.
-
-    Args:
-        adata: an AnnData object.
-        layer: the layer to operate on. Defaults to DKM.X_LAYER.
-    """
-
-    mat = DKM.select_layer_data(adata, layer, copy=False)
-    if issparse(mat):
-        if is_integer_arr(mat.data):
-            mat = mat.asfptype()
-            DKM.set_layer_data(adata, layer, mat)
-        _log2_inplace(mat.data)
-    else:
-        mat = mat.astype(np.float)
-        _log2_inplace(mat)
-
-
-def log_inplace(adata: AnnData, layer: str = DKM.X_LAYER) -> None:
-    """Calculate the natural logarithm `log(exp(x)) = x` for a layer of an AnnData object inplace.
-
-    Args:
-        adata: an AnnData object.
-        layer: the layer to operate on. Defaults to DKM.X_LAYER.
-    """
-
-    mat = DKM.select_layer_data(adata, layer, copy=False)
-    if issparse(mat):
-        if is_integer_arr(mat.data):
-            mat = mat.asfptype()
-            DKM.set_layer_data(adata, layer, mat)
-        _log_inplace(mat.data)
-    else:
-        mat = mat.astype(np.float)
-        _log_inplace(mat)
-
-
-def Freeman_Tukey_inplace(adata: AnnData, layer: str = DKM.X_LAYER) -> None:
-    """Calculate Freeman-Tukey transform for a layer of an AnnData object inplace.
-
-    Args:
-        adata: an AnnData object.
-        layer: the layer to operate on. Defaults to DKM.X_LAYER.
-    """
-    mat = DKM.select_layer_data(adata, layer, copy=False)
-    if issparse(mat):
-        if is_integer_arr(mat.data):
-            mat = mat.asfptype()
-            DKM.set_layer_data(adata, layer, mat)
-        _Freeman_Tukey(mat.data)
-    else:
-        mat = mat.astype(np.float)
-        _Freeman_Tukey(mat)
-
-    mat.data -= 1
-    DKM.set_layer_data(adata, layer, mat)
 
 
 def is_nonnegative(mat: Union[np.ndarray, spmatrix, list]) -> bool:
