+++ dynamo/tools/clustering.py
@@ -13,9 +13,9 @@ from scipy.sparse import csr_matrix, issparse
 from ..configuration import DKM
 from ..dynamo_logger import main_info
 from ..preprocessing.normalization import normalize
-from ..preprocessing.preprocessor_utils import log1p
 from ..preprocessing.QC import filter_genes_by_outliers as filter_genes
 from ..preprocessing.pca import pca
+from ..preprocessing.transform import log1p
 from ..utils import LoggerManager, copy_adata
 from .connectivity import _gen_neighbor_keys, neighbors
 from .utils import update_dict
